requirements.txt
click==8.1.7
huggingface_hub==0.21.0
imageio==2.34.2
matplotlib==3.10.1
numpy==2.2.3
opencv_contrib_python==4.11.0.86
opencv_python==4.9.0.80
pandas==2.2.3
Pillow==9.4.0
scipy==1.13.0
seaborn==0.13.2
scikit-image
tensorboardX==2.6.2.2
timm==1.0.15
torch
torchvision
tqdm==4.66.2
